requirements.txt
annotated-types==0.7.0
appnope==0.1.4
argparse==1.4.0
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gitdb==4.0.11
gitpython==3.1.43
idna==3.10
imageio==2.36.1
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy-loader==0.4
markupsafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
opt-einsum==3.4.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.3.6
prompt-toolkit==3.0.48
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.4
pydantic-core==2.27.2
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pyyaml==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-image==0.25.0
scipy==1.14.1
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
stack_data==0.6.3
sympy==1.13.1
tensorly==0.9.0
tensorly-torch==0.5.0
tifffile==2024.12.12
torch==2.5.1
torch-harmonics==0.7.3
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wandb==0.19.1
wcwidth==0.2.13
wheel==0.44.0
zipp==3.21.0

